adding unbinned likelihood. Added a file to store polynomial info in order to reduce clutter.
Correctly added exposure to background fitting
