+++ threeML/plugins/OGIP/eventlist.py
@@ -17,11 +17,9 @@ from threeML.config.config import threeML_config
 from threeML.exceptions.custom_exceptions import custom_warnings
 from threeML.io.progress_bar import progress_bar
 
-from pha import PHAContainer
+from event_polynomial import polyfit, unbinned_polyfit
 
-# testing
-import multiprocessing
-import joblib
+from pha import PHAContainer
 
 
 class ReducingNumberOfThreads(Warning):
@@ -35,6 +33,7 @@ class ReducingNumberOfSteps(Warning):
 # find out how many splits we need to make
 def ceildiv(a, b):
     return -(-a // b)
+
 
 class EventList(object):
     def __init__(self, arrival_times, energies, n_channels, start_time=None, stop_time=None, dead_time=None,
@@ -257,6 +256,17 @@ class EventList(object):
     poly_order = property(___get_poly_order, ___set_poly_order,
                           doc="Get or set the polynomial order")
 
+    def _exposure_over_interval(self, tmin, tmax):
+        """ calculate the exposure over a given interval  """
+
+        mask = np.logical_and(self._arrival_times >= tmin, self._arrival_times <= tmax)
+
+        interval_deadtime = self._dead_time[mask]
+
+        return (tmax - tmin) - interval_deadtime.sum()
+
+
+
     def set_polynomial_fit_interval(self, *time_intervals_spec):
         """Set the time interval to fit the background.
         Multiple intervals can be input as separate arguments
@@ -362,10 +372,10 @@ class EventList(object):
 
         display(info_df)
 
-    def _fit_global_and_determine_optimum_grade(self, cnts, bins):
+    def _fit_global_and_determine_optimum_grade(self, cnts, bins, exposure):
         # Fit the sum of all the channels to determine the optimal polynomial
         # grade
-        Nintervals = len(bins)
+
 
         # y                         = []
         # for i in range(Nintervals):
@@ -384,12 +394,12 @@ class EventList(object):
         log_likelihoods = []
 
         for grade in range(min_grade, max_grade + 1):
-            polynomial, log_like = _polyfit(bins, cnts, grade)
+            polynomial, log_like = polyfit(bins, cnts, grade, exposure)
 
             log_likelihoods.append(log_like)
 
         # Found the best one
-        deltaLoglike = np.array(map(lambda x: 2 * (x[0] - x[1]), zip(log_likelihoods[:-1], log_likelihoods[1:])))
+        delta_loglike = np.array(map(lambda x: 2 * (x[0] - x[1]), zip(log_likelihoods[:-1], log_likelihoods[1:])))
 
         # print("\ndelta log-likelihoods:")
 
@@ -398,34 +408,33 @@ class EventList(object):
 
         # print("")
 
-        deltaThreshold = 9.0
+        delta_threshold = 9.0
 
-        mask = (deltaLoglike >= deltaThreshold)
+        mask = (delta_loglike >= delta_threshold)
 
         if (len(mask.nonzero()[0]) == 0):
 
             # best grade is zero!
-            bestGrade = 0
+            best_grade = 0
 
         else:
 
-            bestGrade = mask.nonzero()[0][-1] + 1
+            best_grade = mask.nonzero()[0][-1] + 1
 
-        return bestGrade
+        return best_grade
 
     def _fit_polynomials(self):
 
         self._poly_fit_exists = True
-        ## Separate everything by energy channel
 
         # Select all the events that are in the background regions
         # and make a mask
 
         all_bkg_masks = []
 
-        for bkgsel in self._poly_time_selections:
-            all_bkg_masks.append(np.logical_and(self._arrival_times >= bkgsel[0],
-                                                self._arrival_times <= bkgsel[1]))
+        for selection in self._poly_time_selections:
+            all_bkg_masks.append(np.logical_and(self._arrival_times >= selection[0],
+                                                self._arrival_times <= selection[1]))
         poly_mask = all_bkg_masks[0]
 
         # If there are multiple masks:
@@ -433,64 +442,78 @@ class EventList(object):
             for mask in all_bkg_masks[1:]:
                 poly_mask = np.logical_or(poly_mask, mask)
 
+        # Select the all the events in the poly selections
+        # We only need to do this once
+
+        total_poly_events = self._arrival_times[poly_mask]
+
+        # For the channel energies we will need to down select again.
+        # We can go ahead and do this to avoid repeated computations
+
+        total_poly_energies = self._energies[poly_mask]
+
+        # This calculation removes the unselected portion of the light curve
+        # so that we are not fitting zero counts. It will be used in the channel calculations
+        # as well
+
+        bin_width = .1  # seconds
+        these_bins = np.arange(self._start_time,
+                               self._stop_time,
+                               bin_width)
+
+        cnts, bins = np.histogram(total_poly_events,
+                                  bins=these_bins)
+
+        # cnts = cnts / bin_width
+
+
+
+
+        # Find the mean time of the bins and calculate the exposure in each bin
+        mean_time = []
+        exposure_per_bin = []
+        for i in xrange(len(bins) - 1):
+            m = np.mean((bins[i], bins[i + 1]))
+            mean_time.append(m)
+
+            exposure_per_bin.append(self._exposure_over_interval(bins[i], bins[i + 1]))
+
+        mean_time = np.array(mean_time)
+
+        exposure_per_bin = np.array(exposure_per_bin)
+
+        # Remove bins with zero counts
+        all_non_zero_mask = []
+
+        for selection in self._poly_time_selections:
+            all_non_zero_mask.append(np.logical_and(mean_time >= selection[0],
+                                                    mean_time <= selection[1]))
+
+        non_zero_mask = all_non_zero_mask[0]
+        if len(all_non_zero_mask) > 1:
+            for mask in all_non_zero_mask[1:]:
+                non_zero_mask = np.logical_or(mask, non_zero_mask)
+
         # Now we will find the the best poly order unless the use specified one
-        # The total cnts (over channels) is binned to 1 sec intervals
+        # The total cnts (over channels) is binned to .1 sec intervals
+        # TODO: make this unbinned!
         if self._user_poly_order == -1:
-            totalbkgevents = self._arrival_times[poly_mask]
-            bin_width = .1
-            cnts, bins = np.histogram(totalbkgevents,
-                                      bins=np.arange(self._start_time,
-                                                     self._stop_time,
-                                                     bin_width))
-
-            cnts = cnts / bin_width
-            # Find the mean time of the bins
-            mean_time = []
-            for i in xrange(len(bins) - 1):
-                m = np.mean((bins[i], bins[i + 1]))
-                mean_time.append(m)
-            mean_time = np.array(mean_time)
-
-            # Remove bins with zero counts
-            all_non_zero_mask = []
-            
-            for bkgsel in self._poly_time_selections:
-                all_non_zero_mask.append(np.logical_and(mean_time >= bkgsel[0],
-                                                        mean_time <= bkgsel[1]))
-
-            non_zero_mask = all_non_zero_mask[0]
-            if len(all_non_zero_mask) > 1:
-                for mask in all_non_zero_mask[1:]:
-                    non_zero_mask = np.logical_or(mask, non_zero_mask)
 
             self._optimal_polynomial_grade = self._fit_global_and_determine_optimum_grade(cnts[non_zero_mask],
-                                                                                          mean_time[non_zero_mask])
+                                                                                          mean_time[non_zero_mask],
+                                                                                          exposure_per_bin[
+                                                                                              non_zero_mask])
 
             print("Auto-determined polynomial order: %d" % self._optimal_polynomial_grade)
-            print
+            print('\n')
 
 
         else:
 
             self._optimal_polynomial_grade = self._user_poly_order
 
-        # Attempting a parallel execution
-
-        # channels = range(self._first_channel, self._n_channels + self._first_channel)
-        # num_cpus = multiprocessing.cpu_count()
-        # polynomials = joblib.Parallel(n_jobs=num_cpus)(
-        #         joblib.delayed(_fit_channel)(chan,
-        #                                      poly_mask,
-        #                                      self._start_time,
-        #                                      self._stop_time,
-        #                                      self._energies,
-        #                                      self._arrival_times,
-        #                                      self._poly_time_selections,
-        #                                      self._optimal_polynomial_grade) for chan in channels)
-
-
-
         channels = range(self._first_channel, self._n_channels + self._first_channel)
+
         # Check whether we are parallelizing or not
 
 
@@ -500,18 +523,31 @@ class EventList(object):
             polynomials = []
 
             with progress_bar(self._n_channels) as p:
-                for chan in channels:
+                for channel in channels:
 
-                    this_polynomial = _fit_channel(chan,
-                                                   poly_mask,
-                                                   self._start_time,
-                                                   self._stop_time,
-                                                   self._energies,
-                                                   self._arrival_times,
-                                                   self._poly_time_selections,
-                                                   self._optimal_polynomial_grade)
+                    channel_mask = total_poly_energies == channel
 
-                    polynomials.append(this_polynomial)
+                    # Mask background events and current channel
+                    # poly_chan_mask = np.logical_and(poly_mask, channel_mask)
+                    # Select the masked events
+
+                    current_events = total_poly_events[channel_mask]
+
+                    # now bin the selected channel counts
+
+                    cnts, bins = np.histogram(current_events,
+                                              bins=these_bins)
+
+                    # cnts = cnts / bin_width
+
+                    # Put data to fit in an x vector and y vector
+
+                    polynomial, _ = polyfit(mean_time[non_zero_mask],
+                                            cnts[non_zero_mask],
+                                            self._optimal_polynomial_grade,
+                                            exposure_per_bin[non_zero_mask])
+
+                    polynomials.append(polynomial)
                     p.increase()
 
 
@@ -535,34 +571,41 @@ class EventList(object):
                 n_engines = int(self._n_channels)
 
                 custom_warnings.warn(
-                    "The number of engines is larger than the number of channels. Using only %s engines."
-                    % n_engines, ReducingNumberOfThreads)
+                        "The number of engines is larger than the number of channels. Using only %s engines."
+                        % n_engines, ReducingNumberOfThreads)
 
-                # Check if the number of steps is divisible by the number
-                # of threads, otherwise issue a warning and make it so
+            chunk_size = ceildiv(self._n_channels, n_engines)
 
-            # Compute the number of splits, i.e., how many lines in the grid for each engine.
-            # (note that this is guaranteed to be an integer number after the previous checks)
-
-            chunksize = ceildiv(self._n_channels, n_engines)
+            # need to remove class ref
+            grade = self._optimal_polynomial_grade
 
             def worker(start_index):
 
                 polynomials = []
-                channel_subset = channels[chunksize * start_index: chunksize * (start_index + 1)]
+                channel_subset = channels[chunk_size * start_index: chunk_size * (start_index + 1)]
 
-                for chan in channel_subset:
+                for channel in channel_subset:
 
-                    this_polynomial = _fit_channel(chan,
-                                                   poly_mask,
-                                                   self._start_time,
-                                                   self._stop_time,
-                                                   self._energies,
-                                                   self._arrival_times,
-                                                   self._poly_time_selections,
-                                                   self._optimal_polynomial_grade)
+                    channel_mask = total_poly_energies == channel
 
-                    polynomials.append(this_polynomial)
+                    # Select the masked events
+
+                    current_events = total_poly_events[channel_mask]
+
+                    # now bin the selected channel counts
+
+                    cnts, _ = np.histogram(current_events, bins=these_bins)
+
+                    # cnts = cnts / bin_width
+
+                    # Put data to fit in an x vector and y vector
+
+                    polynomial, _ = polyfit(mean_time[non_zero_mask],
+                                            cnts[non_zero_mask],
+                                            grade,
+                                            exposure_per_bin[non_zero_mask])
+
+                    polynomials.append(polynomial)
 
                 return polynomials
 
@@ -576,7 +619,7 @@ class EventList(object):
 
             amr = lview.map_async(worker, range(n_engines))
 
-            #client.wait_watching_progress(amr, 10)
+            # client.wait_watching_progress(amr, 10)
 
             print("\n")
 
@@ -593,510 +636,6 @@ class EventList(object):
         self._polynomials = polynomials
 
 
-def _fit_channel(channel, poly_mask, start_time, stop_time, energies, arrival_times, poly_selections, grade):
-    """ Fit each channel of the data. Function to allow parallel execution
-
-    Args:
-        channel:
-        poly_mask:
-        start_time:
-        stop_time:
-        energies:
-        arrival_times:
-        poly_selections:
-        grade:
-    """
-    # index all events for this channel and select them
-    channel_mask = energies == channel
-
-    # Mask background events and current channel
-    poly_chan_mask = np.logical_and(poly_mask, channel_mask)
-    # Select the masked events
-    current_events = arrival_times[poly_chan_mask]
-
-    bin_width = .1
-
-    cnts, bins = np.histogram(current_events,
-                              bins=np.arange(start_time,
-                                             stop_time,
-                                             bin_width))
-
-    cnts = cnts / bin_width
-
-    # Find the mean time of the bins
-    mean_time = []
-    for i in xrange(len(bins) - 1):
-        m = np.mean((bins[i], bins[i + 1]))
-        mean_time.append(m)
-    mean_time = np.array(mean_time)
-
-    # Remove bins with zero counts
-    all_non_zero_mask = []
-    for bkgsel in poly_selections:
-        all_non_zero_mask.append(np.logical_and(mean_time >= bkgsel[0],
-
-                                                mean_time <= bkgsel[1]))
-
-    non_zero_mask = all_non_zero_mask[0]
-    if len(all_non_zero_mask) > 1:
-        for mask in all_non_zero_mask[1:]:
-            non_zero_mask = np.logical_or(mask, non_zero_mask)
-
-    # Put data to fit in an x vector and y vector
-    polynomial, _ = _polyfit(mean_time[non_zero_mask],
-                             cnts[non_zero_mask],
-                             grade)
-
-    return polynomial
-
-
-def _polyfit(x, y, grade):
-    """ funtion to fit a polynomial to event data. not a member to allow parallel computation """
-    test = False
-
-    # Check that we have enough counts to perform the fit, otherwise
-    # return a "zero polynomial"
-    non_zero_mask = (y > 0)
-    n_non_zero = len(non_zero_mask.nonzero()[0])
-    if n_non_zero == 0:
-        # No data, nothing to do!
-        return Polynomial([0.0]), 0.0
-    pass
-
-    # Compute an initial guess for the polynomial parameters,
-    # with a least-square fit (with weight=1) using SVD (extremely robust):
-    # (note that polyfit returns the coefficient starting from the maximum grade,
-    # thus we need to reverse the order)
-    if test:
-        print("  Initial estimate with SVD..."),
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-
-        initial_guess = np.polyfit(x, y, grade)
-    pass
-    initial_guess = initial_guess[::-1]
-    if (test):
-        print("  done -> %s" % (initial_guess))
-
-    polynomial = Polynomial(initial_guess)
-
-    # Check that the solution found is meaningful (i.e., definite positive
-    # in the interval of interest)
-    M = polynomial(x)
-
-    negative_mask = (M < 0)
-
-    if len(negative_mask.nonzero()[0]) > 0:
-        # Least square fit failed to converge to a meaningful solution
-        # Reset the initialGuess to reasonable value
-        initial_guess[0] = np.mean(y)
-        meanx = np.mean(x)
-        initial_guess = map(lambda x: abs(x[1]) / pow(meanx, x[0]), enumerate(initial_guess))
-
-    # Improve the solution using a logLikelihood statistic (Cash statistic)
-    log_likelihood = PolyLogLikelihood(x, y, polynomial)
-
-    # Check that we have enough non-empty bins to fit this grade of polynomial,
-    # otherwise lower the grade
-    dof = n_non_zero - (grade + 1)
-    if test:
-        print("Effective dof: %s" % (dof))
-    if dof <= 2:
-        # Fit is poorly or ill-conditioned, have to reduce the number of parameters
-        while (dof < 2 and len(initial_guess) > 1):
-            initial_guess = initial_guess[:-1]
-            polynomial = Polynomial(initial_guess)
-            log_likelihood = PolyLogLikelihood(x, y, polynomial)
-        pass
-    pass
-
-    # Try to improve the fit with the log-likelihood
-    # try:
-    if (1 == 1):
-        final_estimate = scipy.optimize.fmin(log_likelihood, initial_guess,
-                                             ftol=1E-5, xtol=1E-5,
-                                             maxiter=1e6, maxfun=1E6,
-                                             disp=False)
-    # except:
-    else:
-        # We shouldn't get here!
-        raise RuntimeError("Fit failed! Try to reduce the degree of the polynomial.")
-    pass
-
-    # Get the value for cstat at the minimum
-    min_log_likelihood = log_likelihood(final_estimate)
-
-    # Update the polynomial with the fitted parameters,
-    # and the relative covariance matrix
-    final_polynomial = Polynomial(final_estimate)
-
-    try:
-        final_polynomial.compute_covariance_matrix(log_likelihood.get_free_derivs)
-    except Exception:
-        raise
-    # if test is defined, compare the results with those obtained with ROOT
-
-
-    return final_polynomial, min_log_likelihood
-
-
-class PolyLogLikelihood(object):
-    '''
-    Implements a Poisson likelihood (i.e., the Cash statistic). Mind that this is not
-    the Castor statistic (Cstat). The difference between the two is a constant given
-    a dataset. I kept Cash instead of Castor to make easier the comparison with ROOT
-    during tests, since ROOT implements the Cash statistic.
-    '''
-
-    def __init__(self, x, y, model, **kwargs):
-        self.x = x
-        self.y = y
-        self.model = model
-        self.parameters = model.get_params()
-
-        # Initialize the exposure to 1.0 (i.e., non-influential)
-        # It will be replaced by the real exposure if the exposure keyword
-        # have been used
-        self.exposure = np.zeros(len(x)) + 1.0
-
-        for key in kwargs.keys():
-            if (key.lower() == "exposure"):
-                self.exposure = np.array(kwargs[key])
-        pass
-
-    pass
-
-    def _evalLogM(self, M):
-        # Evaluate the logarithm with protection for negative or small
-        # numbers, using a smooth linear extrapolation (better than just a sharp
-        # cutoff)
-        tiny = np.float64(np.finfo(M[0]).tiny)
-
-        nontinyMask = (M > 2.0 * tiny)
-        tinyMask = np.logical_not(nontinyMask)
-
-        if (len(tinyMask.nonzero()[0]) > 0):
-            logM = np.zeros(len(M))
-            logM[tinyMask] = np.abs(M[tinyMask]) / tiny + np.log(tiny) - 1
-            logM[nontinyMask] = np.log(M[nontinyMask])
-        else:
-            logM = np.log(M)
-        return logM
-
-    pass
-
-    def __call__(self, parameters):
-        '''
-          Evaluate the Cash statistic for the given set of parameters
-        '''
-
-        # Compute the values for the model given this set of parameters
-        self.model.set_params(parameters)
-        M = self.model(self.x) * self.exposure
-        Mfixed, tiny = self._fix_precision(M)
-
-        # Replace negative values for the model (impossible in the Poisson context)
-        # with zero
-        negativeMask = (M < 0)
-        if (len(negativeMask.nonzero()[0]) > 0):
-            M[negativeMask] = 0.0
-        pass
-
-        # Poisson loglikelihood statistic (Cash) is:
-        # L = Sum ( M_i - D_i * log(M_i))
-
-        logM = self._evalLogM(M)
-
-        # Evaluate v_i = D_i * log(M_i): if D_i = 0 then the product is zero
-        # whatever value has log(M_i). Thus, initialize the whole vector v = {v_i}
-        # to zero, then overwrite the elements corresponding to D_i > 0
-        d_times_logM = np.zeros(len(self.y))
-        nonzeroMask = (self.y > 0)
-        d_times_logM[nonzeroMask] = self.y[nonzeroMask] * logM[nonzeroMask]
-
-        logLikelihood = np.sum(Mfixed - d_times_logM)
-
-        return logLikelihood
-
-    pass
-
-    def _fix_precision(self, v):
-        '''
-          Round extremely small number inside v to the smallest usable
-          number of the type corresponding to v. This is to avoid warnings
-          and errors like underflows or overflows in math operations.
-        '''
-        tiny = np.float64(np.finfo(v[0]).tiny)
-        zeroMask = (np.abs(v) <= tiny)
-        if (len(zeroMask.nonzero()[0]) > 0):
-            v[zeroMask] = np.sign(v[zeroMask]) * tiny
-
-        return v, tiny
-
-    pass
-
-    def get_free_derivs(self, parameters=None):
-        '''
-        Return the gradient of the logLikelihood for a given set of parameters (or the current
-        defined one, if parameters=None)
-        '''
-        # The derivative of the logLikelihood statistic respect to parameter p is:
-        # dC / dp = Sum [ (dM/dp)_i - D_i/M_i (dM/dp)_i]
-
-        # Get the number of parameters and initialize the gradient to 0
-        Nfree = self.model.getNumFreeParams()
-        derivs = np.zeros(Nfree)
-
-        # Set the parameters, if a new set has been provided
-        if (parameters is not None):
-            self.model.set_params(parameters)
-        pass
-
-        # Get the gradient of the model respect to the parameters
-        modelDerivs = self.model.get_free_derivs(self.x) * self.exposure
-        # Get the model
-        M = self.model(self.x) * self.exposure
-
-        M, tinyM = self._fix_precision(M)
-
-        # Compute y_divided_M = y/M: inizialize y_divided_M to zero
-        # and then overwrite the elements for which y > 0. This is to avoid
-        # possible underflow and overflow due to the finite precision of the
-        # computer
-        y_divided_M = np.zeros(len(self.y))
-        nonzero = (self.y > 0)
-        y_divided_M[nonzero] = self.y[nonzero] / M[nonzero]
-
-        for p in range(Nfree):
-            thisModelDerivs, tinyMd = self._fix_precision(modelDerivs[p])
-            derivs[p] = np.sum(thisModelDerivs * (1.0 - y_divided_M))
-        pass
-
-        return derivs
-
-    pass
-
-
-pass
-
-
-class Polynomial(object):
-    def __init__(self, params):
-        self.params = params
-        self.degree = len(params) - 1
-
-        # Build an empty covariance matrix
-        self.cov_matrix = np.zeros([self.degree + 1, self.degree + 1])
-
-    pass
-
-    def horner(self, x):
-        """A function that implements the Horner Scheme for evaluating a
-        polynomial of coefficients *args in x."""
-        result = 0
-        for coefficient in self.params[::-1]:
-            result = result * x + coefficient
-        return result
-
-    pass
-
-    def __call__(self, x):
-        return self.horner(x)
-
-    pass
-
-    def __str__(self):
-        # This is call by the print() command
-        # Print results
-        output = "\n------------------------------------------------------------"
-        output += '\n| {0:^10} | {1:^20} | {2:^20} |'.format("COEFF", "VALUE", "ERROR")
-        output += "\n|-----------------------------------------------------------"
-        for i, parValue in enumerate(self.get_params()):
-            output += '\n| {0:<10d} | {1:20.5g} | {2:20.5g} |'.format(i, parValue, math.sqrt(self.cov_matrix[i, i]))
-        pass
-        output += "\n------------------------------------------------------------"
-
-        return output
-
-    pass
-
-    def set_params(self, parameters):
-        self.params = parameters
-
-    pass
-
-    def get_params(self):
-        return self.params
-
-    pass
-
-    def getNumFreeParams(self):
-        return self.degree + 1
-
-    pass
-
-    def get_free_derivs(self, x):
-        Npar = self.degree + 1
-        freeDerivs = []
-        for i in range(Npar):
-            freeDerivs.append(map(lambda xx: pow(xx, i), x))
-        pass
-        return np.array(freeDerivs)
-
-    pass
-
-    def compute_covariance_matrix(self, statisticGradient):
-        self.cov_matrix = compute_covariance_matrix(statisticGradient, self.params)
-        # Check that the covariance matrix is positive-defined
-        negativeElements = (np.matrix.diagonal(self.cov_matrix) < 0)
-        if (len(negativeElements.nonzero()[0]) > 0):
-            raise RuntimeError(
-                    "Negative element in the diagonal of the covariance matrix. Try to reduce the polynomial grade.")
-
-    pass
-
-    def get_covariance_matrix(self):
-        return self.cov_matrix
-
-    pass
-
-    def integral(self, xmin, xmax):
-        '''
-        Evaluate the integral of the polynomial between xmin and xmax
-        '''
-        integralCoeff = [0]
-        integralCoeff.extend(map(lambda i: self.params[i - 1] / float(i), range(1, self.degree + 1 + 1)))
-
-        integralPolynomial = Polynomial(integralCoeff)
-
-        return integralPolynomial(xmax) - integralPolynomial(xmin)
-
-    pass
-
-    def integral_error(self, xmin, xmax):
-        # Based on http://root.cern.ch/root/html/tutorials/fit/ErrorIntegral.C.html
-
-        # Set the weights
-        i_plus_1 = np.array(range(1, self.degree + 1 + 1), 'd')
-
-        def evalBasis(x):
-            return (1 / i_plus_1) * pow(x, i_plus_1)
-
-        c = evalBasis(xmax) - evalBasis(xmin)
-
-        # Compute the error on the integral
-        err2 = 0.0
-        nPar = self.degree + 1
-        parCov = self.get_covariance_matrix()
-        for i in range(nPar):
-            s = 0.0
-            for j in range(nPar):
-                s += parCov[i, j] * c[j]
-            pass
-            err2 += c[i] * s
-        pass
-
-        return math.sqrt(err2)
-
-    pass
-
-
-pass
-
-
-def compute_covariance_matrix(grad, par, full_output=False,
-                              init_step=0.01, min_step=1e-12, max_step=1, max_iters=50,
-                              target=0.1, min_func=1e-7, max_func=4):
-    """Perform finite differences on the _analytic_ gradient provided by user to calculate hessian/covariance matrix.
-
-    Positional args:
-        grad                : a function to return a gradient
-        par                 : vector of parameters (should be function minimum for covariance matrix calculation)
-
-    Keyword args:
-
-        full_output [False] : if True, return information about convergence, else just the covariance matrix
-        init_step   [1e-3]  : initial step size (0.04 ~ 10% in log10 space); can be a scalar or vector
-        min_step    [1e-6]  : the minimum step size to take in parameter space
-        max_step    [1]     : the maximum step size to take in parameter sapce
-        max_iters   [5]     : maximum number of iterations to attempt to converge on a good step size
-        target      [0.5]   : the target change in the function value for step size
-        min_func    [1e-4]  : the minimum allowable change in (abs) function value to accept for convergence
-        max_func    [4]     : the maximum allowable change in (abs) function value to accept for convergence
-    """
-
-    nparams = len(par)
-    step_size = np.ones(nparams) * init_step
-    step_size = np.maximum(step_size, min_step * 1.1)
-    step_size = np.minimum(step_size, max_step * 0.9)
-    hess = np.zeros([nparams, nparams])
-    min_flags = np.asarray([False] * nparams)
-    max_flags = np.asarray([False] * nparams)
-
-    def revised_step(delta_f, current_step, index):
-        if (current_step == max_step):
-            max_flags[i] = True
-            return True, 0
-
-        elif (current_step == min_step):
-            min_flags[i] = True
-            return True, 0
-
-        else:
-            adf = abs(delta_f)
-            if adf < 1e-8:
-                # need to address a step size that results in a likelihood change that's too
-                # small compared to precision
-                pass
-
-            if (adf < min_func) or (adf > max_func):
-                new_step = current_step / (adf / target)
-                new_step = min(new_step, max_step)
-                new_step = max(new_step, min_step)
-                return False, new_step
-            else:
-                return True, 0
-
-    iters = np.zeros(nparams)
-    for i in xrange(nparams):
-        converged = False
-
-        for j in xrange(max_iters):
-            iters[i] += 1
-
-            di = step_size[i]
-            par[i] += di
-            g_up = grad(par)
-
-            par[i] -= 2 * di
-            g_dn = grad(par)
-
-            par[i] += di
-
-            delta_f = (g_up - g_dn)[i]
-
-            converged, new_step = revised_step(delta_f, di, i)
-            # print 'Parameter %d -- Iteration %d -- Step size: %.2e -- delta: %.2e'%(i,j,di,delta_f)
-
-            if converged:
-                break
-            else:
-                step_size[i] = new_step
-        pass
-
-        hess[i, :] = (g_up - g_dn) / (2 * di)  # central difference
-
-        if not converged:
-            print 'Warning: step size for parameter %d (%.2g) did not result in convergence.' % (i, di)
-    try:
-        cov = np.linalg.inv(hess)
-    except:
-        print 'Error inverting hessian.'
-        raise Exception('Error inverting hessian')
-    if full_output:
-        return cov, hess, step_size, iters, min_flags, max_flags
-    else:
-        return cov
 
 
 def intervals_overlap(tmin, tmax):
